Added the code for k-fold cross-validation 3_1_kfold_train) - this is to run and log the results of cross validation experiments. The following changes were made to incorperate this:
EarlyStopping.py added to ensure that there is a stopping condition for each training run - it isn't feasible to run each experiment for 12+ hours.
Early stopping is added to the training loop found in train_helper.
Added early_stopping and early_stopping_trheshold options to the config file to allow the early stopping to be customised. These options are supplied to the training functions, namely train_resnet and train_helper which are found in utils_models.py.

To facilitate creating subsets of datasets for each cross validation the following changes were made to custom_dataset.py:

SubsetTransform class added to allow transforms to be applied to the subset classes (the subsets for the training and validation classes).
create_image_datasets was added to allow for the one folder containing all the files to create two subset datasets for training and validation therefore facilitating k-fold cross-validation.