Fixing the kfold logs to be the correct destination.
Printing the fold number and validation subjects in terminal when 3_1_kfold_train.py is unning.

Correcting the EarlyStopper so that the indicator is stored as opposed to checking if it is TRUE (= instead of ==)

Adding the accuracy to the EarlyStopper so it can be recoreded.

Removing the need to calculate the statistics in the TRacker when logging the config file.