Adding an explainer for modifying the custome dataset for debugging in custom_dataset.py
utils_model.py: Adding a val_logger function which stores the file information, outputs, predictions etc in a csv which can be post-processed. Additionally changed how the weights of the model are stored so they give an indiation of the date

utils.py: (also changed the order so its clear that the identifier for the weights and the logs are in the format yearmonthday).